chore: move shader manager and metal wrapper to host/
