+++ mopro-msm/src/msm/metal_msm/tests/bigint/bigint_add_wide.rs
@@ -1,5 +1,5 @@
+use crate::msm::metal_msm::host::metal_wrapper::*;
 use crate::msm::metal_msm::utils::limbs_conversion::GenericLimbConversion;
-use crate::msm::metal_msm::utils::metal_wrapper::*;
 
 use ark_ff::{BigInt, BigInteger, UniformRand};
 use ark_std::rand;
