+++ mopro-msm/src/msm/metal_msm/utils/mod.rs
@@ -1,6 +1,4 @@
 pub mod barrett_params;
 pub mod limbs_conversion;
-pub mod metal_wrapper;
 pub mod mont_params;
 pub mod mont_reduction;
-pub mod shader_manager;
